Don't fail when tagging the docker image
The use case here is a workflow rerun when a step within the deploy
workflow failed. Once the tag is added to the image, subsequent runs
would fail as AWS throws an error at the put-image command when the tag
to add already exists on the image.

I didn't find any option on the put-image command to silent this sort of
operation, neither did I find a command to list current image tags (in
the docker sense, not aws resource tags).