chore: add working directory input to node-setup
I haven't yet put the finger on the documentation officially stating
that delegating steps to a custom action do not use the
`working-directory` setting defined in the "caller" workflow. In this
instance, a project using this action in one of its workflows was
setting a default working directory, which was a sub-directory of a git
repository. Then, this action was installing the dependencies, but in
the repository root, which was sort of unexpected to the client
workflow.

The only directly relevant forum thread I found is below.

See: [Sort of related StackOverflow thread](https://stackoverflow.com/questions/58139175/running-actions-in-another-directory)